delete definePlugin.ts
